Use Int instead of Short to report length of data in hfiles (#16586)
(sapling split of 90112bada9b3b0a3cc2cdb630b93198ccec8f534)

+++ src/jvm/io/fsq/hfile/reader/concrete/mmap/Bytes.scala
@@ -37,11 +37,11 @@ object Bytes extends Logger {
     sb.toString
   }
 
-  def readVarlenByteSeqOffsets(mbb: ByteBuffer): (Short, Short) = {
+  def readVarlenByteSeqOffsets(mbb: ByteBuffer): (Int, Int) = {
     val start = mbb.position
-    val size = getVInt(mbb).toShort
+    val size = getVInt(mbb)
     assert(size > 0)
-    val byteSeqOffset = (mbb.position - start).toShort
+    val byteSeqOffset = (mbb.position - start)
     mbb.position(mbb.position + size)
     (byteSeqOffset, size)
   }
